Moved synax APIs to .\CodeAnalysis\Syntax folder
