+++ mc/CodeAnalysis/Syntax/Lexer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace mc.CodeAnalysis
+namespace mc.CodeAnalysis.Syntax
 {
     internal class Lexer
     {
