+++ mc/CodeAnalysis/Syntax/ParenthesizedExpressionSyntax.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace mc.CodeAnalysis
+namespace mc.CodeAnalysis.Syntax
 {
     public sealed class ParenthesizedExpressionSyntax : ExpressionSyntax
     {
