+++ mc/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,4 +1,4 @@
-﻿namespace mc.CodeAnalysis
+﻿namespace mc.CodeAnalysis.Syntax
 {
     public enum SyntaxKind
     {
