+++ mc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace mc.CodeAnalysis
+namespace mc.CodeAnalysis.Syntax
 {
     public abstract class SyntaxNode
     {
